add flush-interval option
When proxying streaming responses, it would not flush the response
writer buffer until full or response is complete. That's still
the default but with this option you can make it flush responses
periodically e.g. once per second.

(modified by ploxiln to be disabled by default, and also to remove
NewReverseProxy() which just calls NewSingleHostReverseProxy())